fix the current version definition code.
For some reason, during the loop to define the current version, the `box.CurrentVersion` attribute was being set to hold the current version in the loop (reference changed on each iteration). This updates the code to use sort.slice to order the versions and take the first (highest version) in the list.

The existing test had the data reorder to prove the failure, and 2 testes added to confirm the behaviour on minor and patch segments of the version number